+++ lib/src/globals.dart
@@ -1,0 +1,8 @@
+/// default config file name
+const String kConfigFileName = 'screenshots.yaml';
+
+/// screenshots environment file name
+const String kEnvFileName = 'env.json';
+
+/// Distinguish device OS.
+enum DeviceType { android, ios }
