+++ test/statusbar_color_test.dart
@@ -1,4 +1,4 @@
-import 'package:screenshots/image_magick.dart';
+import 'package:screenshots/src/image_magick.dart' as im;
 import 'package:test/test.dart';
 
 main() {
@@ -6,7 +6,7 @@ main() {
     final imagePath = './test/resources/0.png';
     final cropSize = '1242x42+0+0';
     // if threshold exceeded select black
-    if (thresholdExceeded(imagePath, cropSize, 0.76)) {
+    if (im.thresholdExceeded(imagePath, cropSize, 0.76)) {
       print('use black statusbar');
     } else {
       print('use white statusbar');
